requirements.txt
flask==2.3.3
requests==2.31.0
waitress==2.1.2
python-dotenv==1.1.0